+++ src/musicbeatstate.hpp
@@ -15,7 +15,7 @@ namespace funkin
     public:
         MusicBeatState(/* args */);
         ~MusicBeatState();
-        virtual void update(double delta);
+        virtual void update(float delta);
         virtual void stepHit();
         virtual void beatHit();
     };
