+++ src/object.cpp
@@ -2,7 +2,7 @@
 #include "game.hpp"
 #include <iostream>
 
-funkin::Object::Object(double x, double y)
+funkin::Object::Object(float x, float y)
 {
     this->position = raylib::Vector2(x, y);
     camera = funkin::Game::defaultCamera;
@@ -12,7 +12,7 @@ funkin::Object::~Object()
     alive = false;
 }
 
-void funkin::Object::update(double delta)
+void funkin::Object::update(float delta)
 {
 }
